requirements.txt
asgiref==3.8.1
Django==4.2.6
djangorestframework==3.15.2
sqlparse==0.5.1
